refactor: merge WorkModelSettings into DataSettings, add collapsible UI and improved demo warning
